requirements.txt
pyperclip>=1.5.24
docopt>=0.6.2
bcrypt>=2.0.0
python-daemon>=2.1.0
fasteners>=0.14.1
